requirements-cloud.txt
-r requirements.txt
google-cloud-datastore
google-cloud-logging
google-cloud-monitoring